[TASK] Compatibility Typo3 8 & Typo3 9

+++ bin/console_clear_cache.sh
@@ -6,11 +6,22 @@ echo ${RED}
 echo "Remove typo3temp Files"
 echo ${NC}
 
-rm -rfv public/typo3temp/*
+if [ -d "public/uploads" ]
+then
+    rm -rfv public/typo3temp/*
+    echo ${RED}
+    echo "Fix Folder Structure"
+    echo ${NC}
+fi
+if [ -d "web/uploads" ]
+then
+    rm -rfv web/typo3temp/*
+    echo ${RED}
+    echo "Fix Folder Structure"
+    echo ${NC}
+fi
 
-echo ${RED}
-echo "Fix Folder Structure"
-echo ${NC}
+
 
 php typo3cms install:fixfolderstructure
 
@@ -34,9 +45,23 @@ curl -insecure http://${local_domain}/flush_cache.php?token=clearcache
 echo ${NC}
 fi
 
-if ! [ -f public/typo3conf/ENABLE_INSTALL_TOOL ]
+if [ -f web/flush_cache.php ]
 then
-touch public/typo3conf/ENABLE_INSTALL_TOOL
+echo ${RED}
+echo "Clear Optcache Now"
+curl -insecure http://${local_domain}/flush_cache.php?token=clearcache
+echo ${NC}
+else
+cp -r vendor/bo/t3toolbox/lib/flush_cache.php web/flush_cache.php
+echo ${RED}
+echo "Clear Optcache Now"
+curl -insecure http://${local_domain}/flush_cache.php?token=clearcache
+echo ${NC}
+fi
+
+if ! [ -f web/typo3conf/ENABLE_INSTALL_TOOL ]
+then
+touch web/typo3conf/ENABLE_INSTALL_TOOL
 fi
 
 echo
